concepts/SmartConnect/task2_llm_chains/src/llm_basics.py
from dotenv import load_dotenv
from typing import Dict, Optional
import os
import openai

# Load environment variables
load_dotenv()

# Configure OpenAI
openai.api_key = os.getenv("OPENAI_API_KEY")

def get_llm_response(prompt: str) -> Optional[str]:
    """
    Get a response from the LLM
    Args:
        prompt: The input prompt
    Returns:
        Response from the LLM or None if failed
    """
    
    try:
        response = openai.ChatCompleoption()
        pass
    except Exception as e:
        pass
    
    pass