concepts/SmartConnect/task2_llm_chains/requirements.txt
# requirements.txt
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.12.0